requirements/base.txt
# TODO: set version for all  packages here
djangorestframework
drf-yasg
django-filter
django-environ
django-cors-headers

psycopg2-binary
redis
gunicorn

# celery
celery
flower
eventlet
django-celery-beat

# for admin
django-admin-rangefilter
django-admin-numeric-filter

# logging
sentry-sdk
